requirements.txt
streamlit
numpy
tensorflow
pillow